+++ lib/refresh_interceptor.dart
@@ -2,5 +2,4 @@ library;
 
 export 'src/refresh_interceptor.dart';
 export 'src/refresh_init.dart';
-export 'src/token_store.dart';
 export 'src/types.dart';